+++ src/app/modules/input-types/components/cedar-input-controlled/cedar-input-controlled.component.ts
@@ -1,18 +1,18 @@
-import {AfterViewInit, Component, Input, OnInit, ViewChild, ViewEncapsulation} from '@angular/core';
-import {FieldComponent} from '../../../shared/models/component/field-component.model';
-import {FormBuilder, FormControl, FormGroup, FormGroupDirective, NgForm, Validators} from '@angular/forms';
-import {ComponentDataService} from '../../../shared/service/component-data.service';
-import {CedarUIComponent} from '../../../shared/models/ui/cedar-ui-component.model';
-import {ActiveComponentRegistryService} from '../../../shared/service/active-component-registry.service';
-import {HandlerContext} from '../../../shared/util/handler-context';
-import {ErrorStateMatcher} from '@angular/material/core';
-import {Observable} from 'rxjs';
-import {debounceTime, distinctUntilChanged, map, startWith, switchMap} from 'rxjs/operators';
-import {IntegratedSearchResponseItem} from '../../../shared/models/rest/integrated-search/integrated-search-response-item';
-import {JsonSchema} from '../../../shared/models/json-schema.model';
-import {ControlledFieldDataService} from '../../../shared/service/controlled-field-data.service';
-import {MessageHandlerService} from '../../../shared/service/message-handler.service';
-import {MatAutocompleteTrigger} from '@angular/material/autocomplete';
+import { AfterViewInit, Component, Input, OnInit, ViewChild, ViewEncapsulation } from '@angular/core';
+import { FieldComponent } from '../../../shared/models/component/field-component.model';
+import { FormBuilder, FormControl, FormGroup, FormGroupDirective, NgForm, Validators } from '@angular/forms';
+import { ComponentDataService } from '../../../shared/service/component-data.service';
+import { CedarUIComponent } from '../../../shared/models/ui/cedar-ui-component.model';
+import { ActiveComponentRegistryService } from '../../../shared/service/active-component-registry.service';
+import { HandlerContext } from '../../../shared/util/handler-context';
+import { ErrorStateMatcher } from '@angular/material/core';
+import { Observable } from 'rxjs';
+import { debounceTime, distinctUntilChanged, map, startWith, switchMap } from 'rxjs/operators';
+import { IntegratedSearchResponseItem } from '../../../shared/models/rest/integrated-search/integrated-search-response-item';
+import { JsonSchema } from '../../../shared/models/json-schema.model';
+import { ControlledFieldDataService } from '../../../shared/service/controlled-field-data.service';
+import { MessageHandlerService } from '../../../shared/service/message-handler.service';
+import { MatAutocompleteTrigger } from '@angular/material/autocomplete';
 export class TextFieldErrorStateMatcher implements ErrorStateMatcher {
   isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
     return !!(control && control.invalid && (control.dirty || control.touched));
@@ -23,7 +23,7 @@ export class TextFieldErrorStateMatcher implements ErrorStateMatcher {
   selector: 'app-cedar-input-controlled',
   templateUrl: './cedar-input-controlled.component.html',
   styleUrls: ['./cedar-input-controlled.component.scss'],
-  encapsulation: ViewEncapsulation.None
+  encapsulation: ViewEncapsulation.None,
 })
 export class CedarInputControlledComponent extends CedarUIComponent implements OnInit, AfterViewInit {
   @ViewChild('autoCompleteInput', { static: false, read: MatAutocompleteTrigger }) trigger: MatAutocompleteTrigger;
@@ -42,7 +42,7 @@ export class CedarInputControlledComponent extends CedarUIComponent implements O
     public cds: ComponentDataService,
     private activeComponentRegistry: ActiveComponentRegistryService,
     private controlledFieldDataService: ControlledFieldDataService,
-    private messageHandlerService: MessageHandlerService
+    private messageHandlerService: MessageHandlerService,
   ) {
     super();
     this.options = fb.group({
@@ -58,38 +58,40 @@ export class CedarInputControlledComponent extends CedarUIComponent implements O
     }
     this.inputValueControl = new FormControl(null, validators);
 
-    if (this.component.valueInfo.defaultValue &&
+    if (
+      this.component.valueInfo.defaultValue &&
       typeof this.component.valueInfo.defaultValue === 'object' &&
-      (this.component.valueInfo.defaultValue as object).hasOwnProperty(JsonSchema.termUri)) {
-      this.setValueUIAndModel(this.component.valueInfo.defaultValue ? this.component.valueInfo.defaultValue[JsonSchema.termUri] : null,
-        this.component.valueInfo.defaultValue ? this.component.valueInfo.defaultValue[JsonSchema.rdfsLabel] : null);
+      Object.hasOwn(this.component.valueInfo.defaultValue as object, JsonSchema.termUri)
+    ) {
+      this.setValueUIAndModel(
+        this.component.valueInfo.defaultValue ? this.component.valueInfo.defaultValue[JsonSchema.termUri] : null,
+        this.component.valueInfo.defaultValue ? this.component.valueInfo.defaultValue[JsonSchema.rdfsLabel] : null,
+      );
     }
 
-    this.filteredOptions = this.inputValueControl.valueChanges
-      .pipe(
-        startWith(''),
-        debounceTime(400),
-        distinctUntilChanged(),
-        switchMap(val => {
-          return this.filter(val || '');
-        })
-      );
+    this.filteredOptions = this.inputValueControl.valueChanges.pipe(
+      startWith(''),
+      debounceTime(400),
+      distinctUntilChanged(),
+      switchMap((val) => {
+        return this.filter(val || '');
+      }),
+    );
   }
 
   ngAfterViewInit(): void {
-    this.trigger.panelClosingActions
-      .subscribe(e => {
-        this.setCurrentValue(this.selectedData?.prefLabel);
-      });
+    this.trigger.panelClosingActions.subscribe((e) => {
+      this.setCurrentValue(this.selectedData?.prefLabel);
+    });
   }
 
   filter(val: string): Observable<IntegratedSearchResponseItem[]> {
     return this.controlledFieldDataService.getData(val, this.component).pipe(
-      map(response => {
+      map((response) => {
         if (response == null) {
           return null;
         } else if (response.collection) {
-          return response.collection.filter(option => {
+          return response.collection.filter((option) => {
             return option.prefLabel.toLowerCase().indexOf(val.toLowerCase()) === 0;
           });
         } else {
@@ -98,7 +100,7 @@ export class CedarInputControlledComponent extends CedarUIComponent implements O
           }
           this.messageHandlerService.errorObject(val, response);
         }
-      })
+      }),
     );
   }
 
@@ -109,7 +111,7 @@ export class CedarInputControlledComponent extends CedarUIComponent implements O
 
   onSelectionChange(option: IntegratedSearchResponseItem): void {
     this.handlerContext.changeControlledValue(this.component, option[JsonSchema.atId], option.prefLabel);
-    if (option){
+    if (option) {
       this.selectedData = option;
     }
   }