+++ src/app/modules/input-types/components/cedar-static-page-break/cedar-static-page-break.component.spec.ts
@@ -8,9 +8,8 @@ describe('CedarStaticPageBreakComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ CedarStaticPageBreakComponent ]
-    })
-    .compileComponents();
+      declarations: [CedarStaticPageBreakComponent],
+    }).compileComponents();
   });
 
   beforeEach(() => {