+++ src/app/modules/input-types/components/cedar-static-youtube/cedar-static-youtube.component.ts
@@ -1,19 +1,18 @@
-import {Component, Input, OnInit, ViewEncapsulation} from '@angular/core';
-import {FormBuilder} from '@angular/forms';
-import {ComponentDataService} from '../../../shared/service/component-data.service';
-import {CedarUIComponent} from '../../../shared/models/ui/cedar-ui-component.model';
-import {ActiveComponentRegistryService} from '../../../shared/service/active-component-registry.service';
-import {HandlerContext} from '../../../shared/util/handler-context';
-import {StaticFieldComponent} from '../../../shared/models/static/static-field-component.model';
+import { Component, Input, OnInit, ViewEncapsulation } from '@angular/core';
+import { FormBuilder } from '@angular/forms';
+import { ComponentDataService } from '../../../shared/service/component-data.service';
+import { CedarUIComponent } from '../../../shared/models/ui/cedar-ui-component.model';
+import { ActiveComponentRegistryService } from '../../../shared/service/active-component-registry.service';
+import { HandlerContext } from '../../../shared/util/handler-context';
+import { StaticFieldComponent } from '../../../shared/models/static/static-field-component.model';
 
 @Component({
   selector: 'app-cedar-static-youtube',
   templateUrl: './cedar-static-youtube.component.html',
   styleUrls: ['./cedar-static-youtube.component.scss'],
-  encapsulation: ViewEncapsulation.None
+  encapsulation: ViewEncapsulation.None,
 })
 export class CedarStaticYoutubeComponent extends CedarUIComponent implements OnInit {
-
   component: StaticFieldComponent;
   @Input() handlerContext: HandlerContext;
   videoHeight: number = null;
@@ -22,13 +21,12 @@ export class CedarStaticYoutubeComponent extends CedarUIComponent implements OnI
   constructor(
     fb: FormBuilder,
     public cds: ComponentDataService,
-    private activeComponentRegistry: ActiveComponentRegistryService
+    private activeComponentRegistry: ActiveComponentRegistryService,
   ) {
     super();
   }
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
   @Input() set componentToRender(componentToRender: StaticFieldComponent) {
     this.component = componentToRender;
@@ -38,5 +36,4 @@ export class CedarStaticYoutubeComponent extends CedarUIComponent implements OnI
   setCurrentValue(currentValue: any): void {
     // DO NOTHING
   }
-
 }