+++ src/app/modules/shared/components/sample-templates/sample-templates.component.spec.ts
@@ -8,9 +8,8 @@ describe('SampleTemplatesComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ SampleTemplatesComponent ]
-    })
-    .compileComponents();
+      declarations: [SampleTemplatesComponent],
+    }).compileComponents();
   });
 
   beforeEach(() => {