+++ src/app/modules/shared/components/source-panels/source-panels.component.ts
@@ -1,14 +1,13 @@
-import {Component, Input, OnInit, ViewEncapsulation} from '@angular/core';
-import {DataContext} from '../../util/data-context';
+import { Component, Input, OnInit, ViewEncapsulation } from '@angular/core';
+import { DataContext } from '../../util/data-context';
 
 @Component({
   selector: 'app-source-panels',
   templateUrl: './source-panels.component.html',
   styleUrls: ['./source-panels.component.scss'],
-  encapsulation: ViewEncapsulation.None
+  encapsulation: ViewEncapsulation.None,
 })
 export class SourcePanelsComponent implements OnInit {
-
   @Input() dataContext: DataContext = null;
 
   @Input() showTemplateRenderingRepresentation: boolean;
@@ -23,14 +22,11 @@ export class SourcePanelsComponent implements OnInit {
   @Input() expandedTemplateRenderingRepresentation = false;
   @Input() expandedMultiInstanceInfo = false;
 
-  constructor() {
-  }
+  constructor() {}
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
   stopPropagation(event): void {
     event.stopPropagation();
   }
-
 }