+++ src/app/modules/shared/handler/data-object-builder.handler.ts
@@ -1,23 +1,22 @@
-import {TemplateComponent} from '../models/template/template-component.model';
-import {CedarComponent} from '../models/component/cedar-component.model';
-import {SingleElementComponent} from '../models/element/single-element-component.model';
-import {MultiElementComponent} from '../models/element/multi-element-component.model';
-import {CedarTemplate} from '../models/template/cedar-template.model';
-import {ElementComponent} from '../models/component/element-component.model';
-import {SingleFieldComponent} from '../models/field/single-field-component.model';
-import {MultiFieldComponent} from '../models/field/multi-field-component.model';
-import {FieldComponent} from '../models/component/field-component.model';
-import {JsonSchema} from '../models/json-schema.model';
+import { TemplateComponent } from '../models/template/template-component.model';
+import { CedarComponent } from '../models/component/cedar-component.model';
+import { SingleElementComponent } from '../models/element/single-element-component.model';
+import { MultiElementComponent } from '../models/element/multi-element-component.model';
+import { CedarTemplate } from '../models/template/cedar-template.model';
+import { ElementComponent } from '../models/component/element-component.model';
+import { SingleFieldComponent } from '../models/field/single-field-component.model';
+import { MultiFieldComponent } from '../models/field/multi-field-component.model';
+import { FieldComponent } from '../models/component/field-component.model';
+import { JsonSchema } from '../models/json-schema.model';
 import * as _ from 'lodash-es';
-import {InstanceExtractData} from '../models/instance-extract-data.model';
-import {InstanceFullData} from '../models/instance-full-data.model';
-import {DataObjectUtil} from '../util/data-object-util';
-import {MultiInstanceObjectHandler} from './multi-instance-object.handler';
-import {CedarInputTemplate} from '../models/cedar-input-template.model';
-import {DataObjectBuildingMode} from '../models/enum/data-object-building-mode.model';
+import { InstanceExtractData } from '../models/instance-extract-data.model';
+import { InstanceFullData } from '../models/instance-full-data.model';
+import { DataObjectUtil } from '../util/data-object-util';
+import { MultiInstanceObjectHandler } from './multi-instance-object.handler';
+import { CedarInputTemplate } from '../models/cedar-input-template.model';
+import { DataObjectBuildingMode } from '../models/enum/data-object-building-mode.model';
 
 export class DataObjectBuilderHandler {
-
   private dataObject: object;
   private dataObjectFull: object;
   private templateJsonObj: object;
@@ -34,13 +33,21 @@ export class DataObjectBuilderHandler {
     return this.getSubTemplate(subTemplate, remainingPath);
   }
 
-  public static buildRecursively(component: CedarComponent, dataObject: InstanceExtractData, templateJsonObj: CedarInputTemplate,
-                                 buildingMode: DataObjectBuildingMode): void {
+  public static buildRecursively(
+    component: CedarComponent,
+    dataObject: InstanceExtractData,
+    templateJsonObj: CedarInputTemplate,
+    buildingMode: DataObjectBuildingMode,
+  ): void {
     let ret = null;
     if (templateJsonObj != null) {
       DataObjectBuilderHandler.addContext(component, dataObject, templateJsonObj, buildingMode);
     }
-    if (component instanceof SingleElementComponent || component instanceof MultiElementComponent || component instanceof CedarTemplate) {
+    if (
+      component instanceof SingleElementComponent ||
+      component instanceof MultiElementComponent ||
+      component instanceof CedarTemplate
+    ) {
       const iterableComponent: ElementComponent = component as ElementComponent;
       const targetName = iterableComponent.name;
       if (component instanceof MultiElementComponent) {
@@ -77,25 +84,28 @@ export class DataObjectBuilderHandler {
           for (let idx = 0; idx < multiField.multiInfo.minItems; idx++) {
             dataObject[targetName].push(DataObjectUtil.getEmptyValueWrapper(subTemplate, buildingMode));
           }
-          const multi = component.choiceInfo.multipleChoice;
-          const values = [];
-          for (const choice of component.choiceInfo.choices) {
-            if (choice.selectedByDefault) {
-              values.push(choice.label);
+          if (component?.choiceInfo?.choices?.length > 0) {
+            const values = [];
+            for (const choice of component.choiceInfo.choices) {
+              if (choice.selectedByDefault) {
+                values.push(choice.label);
+              }
             }
+            dataObject[targetName] = DataObjectUtil.getMultiValueWrapper(subTemplate, buildingMode, values);
           }
-          dataObject[targetName] = DataObjectUtil.getMultiValueWrapper(subTemplate, buildingMode, values);
         }
       } else {
         const subTemplate = DataObjectUtil.getSafeSubTemplate(templateJsonObj, targetName);
         dataObject[targetName] = DataObjectUtil.getEmptyValueWrapper(subTemplate, buildingMode);
-        let value = null;
-        for (const choice of component.choiceInfo.choices) {
-          if (choice.selectedByDefault) {
-            value = choice.label;
+        if (component?.choiceInfo?.choices?.length > 0) {
+          let value = null;
+          for (const choice of component.choiceInfo.choices) {
+            if (choice.selectedByDefault) {
+              value = choice.label;
+            }
           }
+          dataObject[targetName] = DataObjectUtil.getSingleValueWrapper(subTemplate, buildingMode, value);
         }
-        dataObject[targetName] = DataObjectUtil.getSingleValueWrapper(subTemplate, buildingMode, value);
       }
       ret = dataObject[targetName];
     }
@@ -131,8 +141,12 @@ export class DataObjectBuilderHandler {
     }
   }
 
-  private static addContext(component: CedarComponent, dataObject: InstanceExtractData, templateJsonObj: CedarInputTemplate,
-                            buildingMode: DataObjectBuildingMode): void {
+  private static addContext(
+    component: CedarComponent,
+    dataObject: InstanceExtractData,
+    templateJsonObj: CedarInputTemplate,
+    buildingMode: DataObjectBuildingMode,
+  ): void {
     if (buildingMode === DataObjectBuildingMode.INCLUDE_CONTEXT) {
       const props = templateJsonObj[JsonSchema.properties];
       const propsContext = props[JsonSchema.atContext];
@@ -149,7 +163,10 @@ export class DataObjectBuilderHandler {
     this.multiInstanceObjectService = multiInstanceObjectService;
   }
 
-  buildNewExtractDataObject(templateRepresentation: TemplateComponent, templateJsonObj: CedarInputTemplate): InstanceExtractData {
+  buildNewExtractDataObject(
+    templateRepresentation: TemplateComponent,
+    templateJsonObj: CedarInputTemplate,
+  ): InstanceExtractData {
     this.templateJsonObj = templateJsonObj;
     this.templateRepresentation = templateRepresentation;
     this.dataObject = new InstanceExtractData();
@@ -157,7 +174,10 @@ export class DataObjectBuilderHandler {
     return this.dataObject;
   }
 
-  buildNewFullDataObject(templateRepresentation: TemplateComponent, templateJsonObj: CedarInputTemplate): InstanceFullData {
+  buildNewFullDataObject(
+    templateRepresentation: TemplateComponent,
+    templateJsonObj: CedarInputTemplate,
+  ): InstanceFullData {
     this.templateJsonObj = templateJsonObj;
     this.templateRepresentation = templateRepresentation;
     this.dataObjectFull = new InstanceFullData();
@@ -165,13 +185,15 @@ export class DataObjectBuilderHandler {
     return this.dataObjectFull;
   }
 
-  private buildNewByIterating(dataObject: InstanceExtractData, templateJsonObj: CedarInputTemplate,
-                              buildingMode: DataObjectBuildingMode): void {
+  private buildNewByIterating(
+    dataObject: InstanceExtractData,
+    templateJsonObj: CedarInputTemplate,
+    buildingMode: DataObjectBuildingMode,
+  ): void {
     if (this.templateRepresentation != null && this.templateRepresentation.children != null) {
       for (const childComponent of this.templateRepresentation.children) {
         DataObjectBuilderHandler.buildRecursively(childComponent, dataObject, templateJsonObj, buildingMode);
       }
     }
   }
-
 }