+++ src/app/modules/shared/models/component/field-component.model.ts
@@ -1,14 +1,13 @@
-import {CedarComponent} from './cedar-component.model';
-import {BasicInfo} from '../info/basic-info.model';
-import {ValueInfo} from '../info/value-info.model';
-import {NumberInfo} from '../info/number-info.model';
-import {ChoiceInfo} from '../info/choice-info.model';
-import {LabelInfo} from '../info/label-info.model';
-import {ControlledInfo} from '../info/controlled-info.model';
-import {MultiInfo} from '../info/multi-info.model';
+import { CedarComponent } from './cedar-component.model';
+import { BasicInfo } from '../info/basic-info.model';
+import { ValueInfo } from '../info/value-info.model';
+import { NumberInfo } from '../info/number-info.model';
+import { ChoiceInfo } from '../info/choice-info.model';
+import { LabelInfo } from '../info/label-info.model';
+import { ControlledInfo } from '../info/controlled-info.model';
+import { MultiInfo } from '../info/multi-info.model';
 
 export interface FieldComponent extends CedarComponent {
-
   basicInfo: BasicInfo;
   valueInfo: ValueInfo;
   numberInfo: NumberInfo;
@@ -19,5 +18,4 @@ export interface FieldComponent extends CedarComponent {
 
   isMulti(): boolean;
   isMultiPage(): boolean;
-
 }