+++ src/app/modules/shared/models/info/choice-info.model.ts
@@ -1,9 +1,7 @@
-import {ChoiceOption} from './choice-option.model';
+import { ChoiceOption } from './choice-option.model';
 
 export class ChoiceInfo {
-
   multipleChoice: boolean;
   // multipleChoice: number;
   choices: ChoiceOption[] = [];
-
 }