+++ src/app/modules/shared/models/info/controlled-info.model.ts
@@ -1,8 +1,6 @@
 export class ControlledInfo {
-
   ontologies: Array<object>;
   valueSets: Array<object>;
   classes: Array<object>;
   branches: Array<object>;
-
 }