+++ src/app/modules/shared/models/rest/integrated-search/integrated-search-request-parameter-object.ts
@@ -1,8 +1,6 @@
-import {IntegratedSearchRequestValueConstraints} from './integrated-search-request-value-constraints';
+import { IntegratedSearchRequestValueConstraints } from './integrated-search-request-value-constraints';
 
 export class IntegratedSearchRequestParameterObject {
-
   valueConstraints: IntegratedSearchRequestValueConstraints = new IntegratedSearchRequestValueConstraints();
   inputText = '';
-
 }