+++ src/app/modules/shared/models/rest/integrated-search/integrated-search-request-value-constraints.ts
@@ -1,8 +1,6 @@
 export class IntegratedSearchRequestValueConstraints {
-
   ontologies: Array<object> = [];
   valueSets: Array<object> = [];
   classes: Array<object> = [];
   branches: Array<object> = [];
-
 }