+++ src/app/modules/shared/models/template/null-template.model.ts
@@ -1,12 +1,10 @@
-import {TemplateComponent} from './template-component.model';
-import {AbstractElementComponent} from '../element/abstract-element-component.model';
-import {CedarComponent} from '../component/cedar-component.model';
+import { TemplateComponent } from './template-component.model';
+import { AbstractElementComponent } from '../element/abstract-element-component.model';
+import { CedarComponent } from '../component/cedar-component.model';
 
 export class NullTemplate extends AbstractElementComponent implements TemplateComponent {
-
   className = 'NullTemplate';
   pageBreakChildren: Array<CedarComponent[]>;
-
 
   hasPageBreaks(): boolean {
     return false;
@@ -19,5 +17,4 @@ export class NullTemplate extends AbstractElementComponent implements TemplateCo
   isMultiPage(): boolean {
     return false;
   }
-
 }