+++ src/app/modules/shared/models/xsd.model.ts
@@ -1,5 +1,4 @@
 export class Xsd {
-
   static int = 'xsd:int';
   static long = 'xsd:long';
   static float = 'xsd:float';
@@ -7,5 +6,4 @@ export class Xsd {
   static date = 'xsd:date';
   static dateTime = 'xsd:dateTime';
   static time = 'xsd:time';
-
 }